refactor the python code
