+++ pre_process.py
@@ -1,10 +1,39 @@
 import nltk
 import re
 
+def main():
+    """
+    Remove English words, numbers and emojis from the raw text (.txt files)
+    """
+    
+    filename = 'data/raw.txt'
+
+    with open(filename, 'r') as f:
+        text = f.read()
+
+    # Tokenize the text using nltk
+    text = nltk.word_tokenize(text)
+
+    # Clean the text
+    text = clean_text(text)
+    text = remove_english_words(text)
+    text = remove_emoji(text)
+
+
+    # save the cleaned text
+    with open('data/cleaned_tamil.txt', 'w') as f:
+        f.write(' '.join(text))
+
+    print("Text cleaned")
+
+
 def clean_text(words):
     # Remove numbers
     words = [re.sub(r'\d+', '', word) for word in words]
+    print("Numbers removed")
+
     return words
+
 
 def remove_english_words(words):
     # Define a regular expression pattern to match English words
@@ -12,6 +41,8 @@ def remove_english_words(words):
 
     # Filter out English words
     non_english_words = [word for word in words if not re.match(english_word_pattern, word)]
+    print("English words removed")
+
     return non_english_words
 
 def remove_emoji(words):
@@ -20,40 +51,10 @@ def remove_emoji(words):
 
     # Filter out emojis
     non_emoji_words = [word for word in words if not re.match(emoji_pattern, word)]
+    print("Emojis removed")
+
     return non_emoji_words
 
-def main():
-    # Sample Tamil text
-    # text = " slkdfj உலகம் test மிகப்பெரிய தடைகள் y not ? அடைந்துள்ளது. இது திரையை ehlp புதுப்பிக்க தடை எடுத்துவிடுகிறது."
-    
-    filename = 'data/raw.txt'
-
-    with open(filename, 'r') as f:
-        print('Reading file: {}'.format(filename))
-        text = f.read()
-        print('File read successfully')
-
-
-
-    # Tokenize the text using nltk
-    text = nltk.word_tokenize(text)
-    print('Tokenization done')
-
-    # Clean the text
-    text = clean_text(text)
-    print('Cleaning done')
-    text = remove_english_words(text)
-    print('English words removed')
-    text = remove_emoji(text)
-    print('Emojis removed')
-
-    # print(text)
-
-    # save the cleaned text
-    with open('data/cleaned_tamil.txt', 'w') as f:
-        print('Writing to file: data/cleaned_tamil.txt')
-        f.write(' '.join(text))
-        print('File written successfully')
 
 if __name__ == '__main__':
     main()
